Enforce ordering by weights
Now that I look at it, weights really don't matter all that much. More
of it is the fact that you've used that dir in the past. The weights
really just seem to be there to differentiate between C:\ProjA\bin\Debug
and C:\ProjB\bin\Debug based on usage.