Made jumpstat queryable
I broke jumpstat into it's own Get-JumpStatus cmdlet to start
maintaining separation of concerns. Once this was done, it was obvious
to implement querying via Get-JumpStatus since it's already returning
objects instead of changing directories.

This is a great step towards scripting this a lot better than autojump
was done. For one, it'll be easier to manipulate the DB (once we expose
-Save). But moreso, this is a great step toward using jumpstat as input
into other commands. I'd love to be able to do

PS> explorer $(jumpstat ju)

And have it jump right to my favorite project :)